worker/app/requirements.txt
Flask==2.2.5
requests==2.31.0
mysql-connector-python==8.2.0
schedule==1.2.0